feat(misc): move create-nx-plugin create script into create-nx-workspace package
