+++ scripts/build.sh
@@ -29,7 +29,7 @@ rm -rf build/packages/angular/bundles/nrwl-angular-testing.umd.min.js.bak
 
 rsync -a --exclude=*.ts packages/ build/packages
 chmod +x build/packages/create-nx-workspace/bin/create-nx-workspace.js
-chmod +x build/packages/nx-plugin/bin/create.js
+chmod +x build/packages/create-nx-workspace/bin/create-nx-plugin.js
 chmod +x build/packages/cli/bin/nx.js
 chmod +x build/packages/tao/index.js
